Added two example files
One of those files shows multiplication of two row major matrices in c. The other file shows multiplication of one row-major and a column-major matrix. The code also measures speed and hopefully shows a great speed improvement when using one column-major matrix. 

This is intended as a showcase on how to use c syntax for matrix multiplication. Note that this is only one option to access a two-dimensional matrix.